relative motion with thorlabs stage

+++ CTF/ctf_analyser.py
@@ -260,7 +260,7 @@ class CTFAnalyser:
         h, w = self.template.shape[:2]
         if np.all(w == np.abs(self.bboxes[:, 0] - self.bboxes[:, 2])) and np.all(h == np.abs(self.bboxes[:, 1] - self.bboxes[:, 3])):
 
-            out_frame = np.zeros((3*h, 5*w, 3))
+            out_frame = np.zeros((3*h, 5*w, 3), dtype=np.uint8)
 
             out_frame[:h, :w] = crop_frame_bbox(frame, self.bboxes[10-1], str(10))
             out_frame[h:2*h, :w] = crop_frame_bbox(frame, self.bboxes[11-1], str(11))
